Update saving-data-to-the-database.liquid
The GraphQL section (for first query in the GUI) took me a second to figure out because it had some code pre-populated in the query section. Maybe mention that to the user in the docs so they can delete it first (because it changed the behavior of the play button).

Someone I know pointed this out whilst going through the getting started. Not too much of a change but I'll let you decide whether it should go through or not. Thanks : )